// src/types/airtable.generated.ts
export interface AirtableActivity {
  id: string;
  createdTime: string;
  fields: {
    [key: string]: any;
    /** formula field (fld6pXq1O74quaBeM) */
    ID: string;
    /** multipleRecordLinks field (fldl9Pzx0SnlSrkXv) */
    'ID Lead'?: string[];
    /** multipleLookupValues field (fldjTRENBzctRrCGL) */
    'Nome Lead': string;
    /** multipleRecordLinks field (fldY0qOsEB266QoZ8) */
    Assegnatario?: string[];
    /** multipleLookupValues field (fld1tvUfoVtr9G9PK) */
    'Nome Assegnatario': string;
    /** formula field (fldMHDDUqDvOOCLTt) */
    Titolo: string;
    /** singleSelect field (fldcR1a25ViHoT3mY) */
    Tipo?: 'Chiamata' | 'WhatsApp' | 'Email' | 'SMS' | 'Consulenza' | 'Follow-up' | 'Altro';
    /** singleSelect field (fldcEU2D7zwwj28bS) */
    Stato?: 'Da Pianificare' | 'Pianificata' | 'In corso' | 'In attesa' | 'Completata' | 'Annullata' | 'Rimandata';
    /** dateTime field (fldPvGAPV9iYGKvtJ) */
    Data?: string;
    /** duration field (fldpBB4UKATOiXLru) */
    'Durata stimata'?: any;
    /** singleSelect field (fldDA3bPMzKloIWa2) */
    Obiettivo?: 'Primo contatto' | 'Qualificazione lead' | 'Presentazione prodotto' | 'Invio preventivo' | 'Follow-up preventivo' | 'Negoziazione' | 'Chiusura ordine' | 'Fissare appuntamento' | 'Confermare appuntamento' | 'Promemoria appuntamento' | 'Consegna prodotto' | 'Assistenza tecnica' | 'Controllo soddisfazione' | 'Upsell Cross-sell' | 'Richiesta recensione';
    /** singleSelect field (fldjihrvpzgl1bc4h) */
    Priorità?: 'Bassa' | 'Media' | 'Alta' | 'Urgente';
    /** multilineText field (fldrcsQBXJAbwXtJX) */
    Note?: string;
    /** singleSelect field (fldaY6kTm1VVtZgGm) */
    Esito?: 'Contatto riuscito' | 'Nessuna risposta' | 'Numero errato' | 'Non disponibile' | 'Non presentato' | 'Molto interessato' | 'Interessato' | 'Poco interessato' | 'Non interessato' | 'Informazioni raccolte' | 'Preventivo richiesto' | 'Preventivo inviato' | 'Appuntamento fissato' | 'Ordine confermato' | 'Opportunità persa' | 'Servizio completato' | 'Problema risolto' | 'Cliente soddisfatto' | 'Recensione ottenuta';
    /** singleSelect field (fldrJg8nLxcJP4hbF) */
    'Prossima azione'?: 'Chiamata' | 'WhatsApp' | 'Email' | 'SMS' | 'Consulenza' | 'Follow-up' | 'Nessuna';
    /** dateTime field (fldS3v6etzoTOExOP) */
    'Data prossima azione'?: string;
    /** multilineText field (fldPcEwR5jCyIhPvC) */
    Allegati?: string;
    /** createdTime field (fldy429iW236noQ8B) */
    'Creato il '?: any;
    /** lastModifiedTime field (fldzEZhPBXa68idd0) */
    'Ultima modifica'?: any;
    /** singleLineText field (fldSDUhq6QLMQPOsz) */
    Order?: string;
  };
}